feat: Add userEvent package.
@testing-library/user-event has a peerDep for @testing-library/dom
but it's a different version than the one that @testing-library/react
uses and brings in as a devDep. Solution based on research on GitHub
shows we can either ignore the warning in pnpm or install the
@testing-library/dom as devDep in the version that @testing-library/react
uses it. I did the latter hoping that in one of the next versions of
@testing-library/user-event this will be resolved.

see
https://github.com/testing-library/user-event/pull/438#issuecomment-680134095
https://github.com/testing-library/user-event/issues/551#issuecomment-930963323