feat(coral): Add initial layout component
The purpose of this is to only illustrate how we would like to
use layout components. It will be easier to find the proper usage
with some real life experience.

Extend render helpers to include a utility, which renders also
MemoryRouter in the wrapper. The router is required when the component
uses e.g. <Link /> components in navigation.