refactoring(coral): Use DataTable action column type for table actions #797
Unifies the usage of the DataTable actions in the approvals views
for more unified experience. In practice, changes from column type "custom" into
"action".

Fixes one corner case:
- the isBeing* check per request did not consider if the request was
loading. This made it "loading" for example in case where the request
failed.
- when request was declined, it was technically possible to click the button again after modal closed.
It now disables the row's decline button as well until the decline request in finished.

Tries to do minimal changes to tests to ease review process. Adds couple more
test cases per table.

Note*: prior to this commit, we hide the action button if it's not
feasible per request, for example: not in CREATED state.
After this change, the button will always visible but will be
disbaled if the status is not CREATED. This is due to the action column
api, it does not support hiding the column content.

Signed-off-by: Roope Karhapää <roope-kar@aiven.io>